qd_scan/requirements.txt
numpy
scipy
matplotlib
pyyaml
pytest
